split json editor into different files
