+++ webapp/js/constants/actionTypes.js
@@ -11,15 +11,6 @@ export const CHANGE_WINDOW = 'CHANGE_WINDOW';
 export const SEARCH_VALUE_CHANGE = 'SEARCH_VALUE_CHANGE';
 export const LOAD_VALUE = 'LOAD_VALUE';
 
-export const SEARCH_JSON = 'SEARCH_JSON';
-export const EXPAND_JSON = 'EXPAND_JSON';
-export const COLLAPSE_JSON = 'COLLAPSE_JSON';
-export const REDO_JSON = 'REDO_JSON';
-export const UNDO_JSON = 'UNDO_JSON';
-export const CHANGE_JSON_MODE = 'CHANGE_JSON_MODE';
-export const FORMAT_JSON = 'FORMAT_JSON';
-export const COMPACT_JSON = 'COMPACT_JSON';
-
 export const FILTER_BY_SEARCH = 'FILTER_BY_SEARCH';
