+++ webapp/js/constants/jsonEditorTypes.js
@@ -1,0 +1,9 @@
+export const SEARCH_JSON = 'SEARCH_JSON';
+export const EXPAND_JSON = 'EXPAND_JSON';
+export const COLLAPSE_JSON = 'COLLAPSE_JSON';
+export const REDO_JSON = 'REDO_JSON';
+export const UNDO_JSON = 'UNDO_JSON';
+export const CHANGE_JSON_MODE = 'CHANGE_JSON_MODE';
+export const FORMAT_JSON = 'FORMAT_JSON';
+export const COMPACT_JSON = 'COMPACT_JSON';
+export const VALUE_CHANGE = 'VALUE_CHANGE';
